pyproject.toml
[project]
name = "assignment0"
version = "1.0.0"
description = "CIS6930 Assignment 0: API data collection + NaviGator AI summary (PokeAPI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["assignment0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
